+++ CHANGES.rst
@@ -8,7 +8,7 @@ Here you can see the full list of changes between each SQLAlchemy-Utils release.
 ^^^^^^^^^^^^^^^^^^^
 
 - Made assert_* functions automatically rollback session
-
+- Fix alias handling in has_unique_index
 
 
 0.27.4 (2014-10-23)